+++ R/plotAnimatedPPGM.R
@@ -1,7 +1,7 @@
 #' @title plotAnimatedPPGM
 #' @description This function creates an animated gif showing the change in modelled suitable habitat through time in geographic space.
 #' @usage plotAnimatedPPGM(envelope, tree, filename="ppgm.gif", which.biovars, 
-#' path="", use.paleoclimate=TRUE, paleoclimateUser=NULL, layerAge=c(0:20))
+#' path="", use.paleoclimate=TRUE, paleoclimateUser=NULL, layerAge=c(0:20), ...)
 #' @param envelope the min and max envelope of each lineage for each time slice
 #' @param tree the phylogeny or multiple phylogenies that show the relationship between species
 #' @param filename desired filename of output
@@ -10,11 +10,12 @@
 #' @param use.paleoclimate if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided
 #' @param paleoclimateUser list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19. (see \code{getBioclimvars()}).
 #' @param layerAge vector with the ages of the paleoclimate dataframes, if using user submitted paleoclimate data
+#' @param ... other parameters to pass to save_gif
 #' @details Requires ImageMagick or GraphicsMagick to be installed on the operating system. This is easy to do if you have macports. Just type sudo port install ImageMagick into terminal.
 #' @return An animated gif of species through time
 #' @author A. Michelle Lawing, Alexandra F. C. Howard, Maria-Aleja Hurtado-Materon
 #' @importFrom utils data
-#' @importFrom animation saveGIF
+#' @importFrom gifski save_gif
 #' @importFrom graphics points
 #' @importFrom grDevices colorRampPalette
 #' @export
@@ -32,8 +33,7 @@
 #' example_getEnvelopes <- getEnvelopes(treedata_min, treedata_max, node_est)
 #' animatedplot <- plotAnimatedPPGM(example_getEnvelopes,tree,which.biovars=1,path=tempdir())}
 
-
-plotAnimatedPPGM<-function(envelope, tree, filename="ppgm.gif", which.biovars, path="", use.paleoclimate=TRUE, paleoclimateUser=NULL, layerAge=c(0:20)){
+plotAnimatedPPGM<-function(envelope, tree, filename="ppgm.gif", which.biovars, path="", use.paleoclimate=TRUE, paleoclimateUser=NULL, layerAge=c(0:20),...){
   #load paleoclimate data: isotopically scaled paleoclimate bioclimate variables for North America
   if(use.paleoclimate) {
     paleoclimate <- paleoclimate #uses paleoclimate data from package
@@ -64,8 +64,9 @@ plotAnimatedPPGM<-function(envelope, tree, filename="ppgm.gif", which.biovars, p
     hld[which(hld==0,arr.ind=TRUE)]=NA
     return(hld)
   })
-  animation::saveGIF(for(i in 1:length(paleoclimate)){
-    plot(paleoclimate[[i]][,2:3],cex=0.5,pch=16,col="lightgray",xlim=c(-180,0),ylim=c(0,90))
-    graphics::points(paleoclimate[[i]][,2:3],cex=0.5,pch=16,col=grDevices::colorRampPalette(c("#FFE5CC", "#FF8000", "#990000"))(length(temp_min[[1]][1,]))[richnesscount[[i]]],xlim=c(-180,0),ylim=c(0,90))
-  },movie.name=filename,outdir=getwd())
+  gifski::save_gif(for(i in 1:length(paleoclimate)){
+    plot(paleoclimate[[i]][,2:3],cex=0.5,pch=16,col="lightgray",xlim=c(min(paleoclimate[[1]][,2])-5,max(paleoclimate[[1]][,2])+5),ylim=c(min(paleoclimate[[1]][,3])-5,max(paleoclimate[[1]][,3])+5))
+    graphics::points(paleoclimate[[i]][,2:3],cex=0.5,pch=16,col=grDevices::colorRampPalette(c("#FFE5CC", "#FF8000", "#990000"))(length(temp_min[[1]][1,]))[richnesscount[[i]]],xlim=c(min(paleoclimate[[1]][,2])-5,max(paleoclimate[[1]][,2])+5),ylim=c(min(paleoclimate[[1]][,3])-5,max(paleoclimate[[1]][,3])+5))
+  },gif_file=filename,...)
 }
+