+++ R/plotAnimatedPPGMMultiPhylo.R
@@ -1,7 +1,8 @@
 #' @title plotAnimatedPPGMMultiPhylo
 #' @description This function creates an animated gif showing the change in modeled suitable habitat through time in geographic space. It requires ImageMagick or GraphicsMagick to be previously installed in the operating system. This is easy to do if you have macports. Just type sudo port install ImageMagick into terminal.
 #' @usage plotAnimatedPPGMMultiPhylo(envelope, tree, filename="ppgm.gif", 
-#' which.biovars, path="", use.paleoclimate=TRUE, paleoclimateUser=NULL, layerAge=c(0:20))
+#' which.biovars, path="", use.paleoclimate=TRUE, 
+#' paleoclimateUser=NULL, layerAge=c(0:20), ...)
 #' @param envelope the min and max envelope of each lineage for each time slice
 #' @param tree the phylogeny or multiple phylogenies that show the relationship between species
 #' @param filename filename of output
@@ -10,11 +11,12 @@
 #' @param use.paleoclimate if left blank, default North America paleoclimate data is used. If FALSE, user submitted paleoclimate must be provided
 #' @param paleoclimateUser list of data frames with paleoclimates, must be dataframes with columns: GlobalID, Longitude, Latitude, bio1, bio2,...,bio19. (see \code{getBioclimvars()}).
 #' @param layerAge vector with the ages of the paleoclimate dataframes, if using user submitted paleoclimate data
+#' @param ... other parameters to pass to save_gif
 #' @details Requires ImageMagick or GraphicsMagick to be installed on the operating system. This is easy to do if you have macports. Just type sudo port install ImageMagick into terminal.
 #' @return An animated gif of species through time
 #' @author A. Michelle Lawing, Alexandra F. C. Howard
 #' @importFrom utils data
-#' @importFrom animation saveGIF
+#' @importFrom gifski save_gif
 #' @importFrom graphics points
 #' @importFrom grDevices colorRampPalette
 #' @importFrom graphics text
@@ -36,8 +38,7 @@
 #' }
 #' animatedplot <- plotAnimatedPPGMMultiPhylo(envelope,sampletrees,which.biovars=1, path=tempdir())}
 
-
-plotAnimatedPPGMMultiPhylo <- function(envelope, tree, filename="ppgm.gif", which.biovars, path="", use.paleoclimate=TRUE, paleoclimateUser=NULL, layerAge=c(0:20)){
+plotAnimatedPPGMMultiPhylo <- function(envelope, tree, filename="ppgm.gif", which.biovars, path="", use.paleoclimate=TRUE, paleoclimateUser=NULL, layerAge=c(0:20),...){
   if(path==""){out=getwd()}
   if(path!=""){out=paste(getwd(),"/",substr(path,1,nchar(path)-1),sep="")}
   #load paleoclimate data
@@ -62,23 +63,25 @@ plotAnimatedPPGMMultiPhylo <- function(envelope, tree, filename="ppgm.gif", whic
       temp<-lapply(1:length(which.biovars),function(j){getTimeSlice(layerAge[[i]],tree[[tr]],envelope[[tr]][,5,j])})
       temp<-t(array(unlist(temp),dim=c(length(unlist(temp[[1]]$edge)),2*length(which.biovars))))
       return(temp)})
+    allhld <- as.list(array(NA,dim=length(paleoclimate)))
     for (j in 1:length(paleoclimate)){
       hld<-array(0,dim=length(paleoclimate[[j]][,1]))
       for(i in 1:length(temp_min[[tr]][[j]][1,])){
         matching <- sapply(1:length(which.biovars),function(x){
           paleoclimate[[j]][,which.biovars[x]+3]>temp_min[[tr]][[j]][1:length(which.biovars)*2,i][x] & 
             paleoclimate[[j]][,which.biovars[x]+3]<temp_max[[tr]][[j]][1:length(which.biovars)*2,i][x]
-          })
+        })
         matching<-which(rowSums(matching)==length(which.biovars),arr.ind=TRUE)
         hld[matching]<-hld[matching]+1
       }
       hld[which(hld==0,arr.ind=TRUE)]=NA
+      allhld[[j]] <-hld
     }
-    richnesscount[[tr]] <-hld
+    richnesscount[[tr]] <- allhld
   }
   oldpar <- par(no.readonly = TRUE)
   on.exit(par(oldpar))
-  animation::saveGIF(for(i in 1:length(paleoclimate)){
+  gifski::save_gif(for(i in 1:length(paleoclimate)){
     richnesscountMean<-rowMeans(array(unlist(lapply(1:length(tree),function(tr) richnesscount[[tr]][[i]])),dim=c(length(richnesscount[[1]][[i]]),length(tree))),na.rm=TRUE)
     lq<-array(unlist(lapply(1:length(tree),function(tr) richnesscount[[tr]][[i]])),dim=c(length(richnesscount[[1]][[i]]),length(tree)))
     richnesscountMIN<-suppressWarnings(unlist(lapply(1:length(lq[,1]),function(r) min(lq[r,],na.rm=TRUE))))
@@ -94,9 +97,11 @@ plotAnimatedPPGMMultiPhylo <- function(envelope, tree, filename="ppgm.gif", whic
     translate<-cbind(scale[,1]-35,scale[,2]+25)
     graphics::points(translate,cex=0.5,pch=16,col="lightgray")
     graphics::points(translate,cex=0.5,pch=16,col=colorRampPalette(c("#FFE5CC", "#FF8000", "#990000"))(length(temp_min[[tr]][[1]][1,]))[richnesscountMAX])
+    graphics::text(-180,90,paste0(i-1,"mya"))
     graphics::text(-160,40,"Min")
     graphics::text(-35,40,"Max")
     graphics::text(-90,85,"Mean")
-    graphics::text(-90,90,"Modeled Richness over a distribution of trees")
-  },movie.name=filename,outdir=out)
+    graphics::text(-90,90,"Modeled Suitable Climate")
+  },gif_file=filename,...)
 }
+