+++ R/sampletrees.R
@@ -1,8 +1,10 @@
 #' Sample of Sceloporus phylogenies
 #' 
-#' A sample of 100 dated phylogenies from Leache & Sites (2010), trimmed for analysis by Lawing et al (2016)
-#' Lawing et al (2016) Including fossils in phylogenetic climate reconstructions: A deep time perspective on the climatic niche evolution and diversification of spiny lizards (Sceloporus)
-#' Leache & Sites (2010) Chromosome evolution and diversification in North American Spiny Lizards (Genus Sceloporus)
+#' A sample of 100 dated phylogenies from Leache & Sites (2009), trimmed for analysis by Lawing et al (2016).
+#' \itemize{
+#'  \item Lawing et al (2016) Including fossils in phylogenetic climate reconstructions: A deep time perspective on the climatic niche evolution and diversification of spiny lizards (Sceloporus) The American Naturalist 188(2): 133-148
+#'  \item Leache & Sites (2009) Chromosome evolution and diversification in North American Spiny Lizards (Genus Sceloporus) Cytogenetic and Genome Research 127(2-4) 166-181
+#'  }
 #' 
 #' @format 'sampletrees'
 #' 100 trees as class multiPhylo