+++ tests/testthat/test-plotAnimatedPPGMMultiPhylo.R
@@ -7,6 +7,5 @@ test_that("animatedppgmMultiPhylo works", {
   testclim <- list(testclim[[1]],testclim[[2]],testclim[[3]])
   #test models
   tBM  <- ppgm(testocc, trees=testtree, model="BM", which.biovars=c(1,4,15), use.paleoclimate=F, paleoclimateUser=testclim, layerAge=c(0:2))
-  expect_snapshot(plotAnimatedPPGMMultiPhylo(tBM$envelope, tree=testtree, which.biovars=c(1,4,15), use.paleoclimate=F, paleoclimateUser=testclim, layerAge=c(0:2)))
+  expect_snapshot_file(plotAnimatedPPGMMultiPhylo(tBM$envelope, tree=testtree, which.biovars=c(1,4,15), use.paleoclimate=F, paleoclimateUser=testclim, layerAge=c(0:2)))
 })
-