+++ tests/testthat/test-ppgm.R
@@ -16,7 +16,7 @@ test_that("ppgm works with original paleoclimate", {
   #test est
   tEs <- ppgm(testocc, trees=testtree, model="estimate", which.biovars=1, use.paleoclimate=F, paleoclimateUser=testclim, bounds=bounds)
   #test fossil
-  tfos <- ppgm(testocc, trees=testtree, fossils=testfos, which.biovars=1, use.paleoclimate=F, paleoclimateUser=testclim)
+  tfos <- ppgm(testocc, trees=testtree, model="BM", fossils=testfos, which.biovars=1, use.paleoclimate=F, paleoclimateUser=testclim)
   #checking outputs
   expect_equal(length(tBM$node_est),length(testtree))
   expect_equal(length(tfos$node_est),length(testtree))
@@ -35,9 +35,9 @@ test_that("ppgm works with new paleoclimate", {
   layerAge <- c(1,4,10)
   bounds <- list(a = c(min = -1, max = 5), delta = c(min = 0, max = 1000))
   #test BM
-  tBM  <- ppgm(testocc, trees=testtree, model="BM" ,which.biovars=1, use.paleoclimate=F, paleoclimateUser=newclim, layerAge=layerAge)
+  tBM  <- ppgm(testocc, trees=testtree, model="BM", which.biovars=1, use.paleoclimate=F, paleoclimateUser=newclim, layerAge=layerAge)
   #test fossil
-  tfos <- ppgm(testocc, trees=testtree, fossils=testfos, which.biovars=1, use.paleoclimate=F, paleoclimateUser=newclim, layerAge=layerAge)
+  tfos <- ppgm(testocc, trees=testtree, model="BM", fossils=testfos, which.biovars=1, use.paleoclimate=F, paleoclimateUser=newclim, layerAge=layerAge)
   #checking outputs
   expect_equal(length(tBM$node_est),length(testtree))
   expect_equal(length(tfos$node_est),length(testtree))